Introduced SolutionMove as base class for moves. Introduced SolutionMoveStops as class to move stops (Move is aliased to this type).
